+++ frontend/src/components/metadata-credits-table.vue
@@ -1,5 +1,8 @@
 <template>
-  <v-simple-table v-if="value && value.length > 0" id="creditsTable">
+  <v-simple-table
+    v-if="sortedCredits && sortedCredits.length > 0"
+    id="creditsTable"
+  >
     <template #default>
       <h2>Credits</h2>
       <table>
@@ -10,7 +13,7 @@
           </tr>
         </thead>
         <tbody>
-          <tr v-for="credit in value" :key="credit.pk">
+          <tr v-for="credit in sortedCredits" :key="credit.pk">
             <td>
               {{ roleName(credit.role.name) }}
             </td>
@@ -36,12 +39,38 @@ export default {
   data() {
     return {};
   },
+  computed: {
+    sortedCredits: function () {
+      /* eslint-disable-next-line array-func/prefer-array-from */
+      return [...this.value].sort(this.creditsCompare);
+    },
+  },
   methods: {
     roleName: function (name) {
       if (name === "CoverArtist") {
         return "Cover Artist";
       }
       return name;
+    },
+    creditSortable: function (credit) {
+      const names = credit.person.name.split(" ");
+      let sortArray = [];
+      if (names) {
+        sortArray.push(names.slice(-1));
+        if (names.length > 1) {
+          sortArray.push(names.slice(0, -1));
+        }
+      }
+      sortArray.push(credit.role.name);
+
+      return sortArray.join(" ");
+    },
+    creditsCompare: function (creditA, creditB) {
+      const creditSortableA = this.creditSortable(creditA);
+      const creditSortableB = this.creditSortable(creditB);
+      if (creditSortableA < creditSortableB) return -1;
+      if (creditSortableA > creditSortableB) return 1;
+      return 0;
     },
   },
 };